src/Search/mu_agent.rs
use std::collections::HashMap;
use std::rc::Rc;
use std::cell::RefCell;
use ndarray::Array1;
use rustc_hash::FxHashMap;

// Define the board size here
const BOARD_SIZE: usize = 8;

// Node structure with optimized data handling
struct Node {
    state: Array1<f64>,
    action: Option<usize>,
    reward: f64,
    value: f64,
    policy: Option<Array1<f64>>,
    visit_count: u32,
    children: FxHashMap<usize, Rc<RefCell<Node>>>,
}

impl Node {
    fn new(state: Array1<f64>) -> Self {
        Node {
            state,
            action: None,
            reward: 0.0,
            value: 0.0,
            policy: None,
            visit_count: 0,
            children: FxHashMap::default(),
        }
    }

    fn expand(&mut self, policy: Array1<f64>, reward: f64, value: f64) {
        self.reward = reward;
        self.value = value;
        self.policy = Some(policy);
        self.visit_count = 1;
        for action in 0..policy.len() {
            self.children.insert(action, Rc::new(RefCell::new(Node::new(Array1::zeros(0)))));
        }
    }

    fn select_child(&self, c_puct: f64) -> (usize, Rc<RefCell<Node>>) {
        let mut best_value = f64::NEG_INFINITY;
        let mut best_action = 0;
        let mut best_child = Rc::new(RefCell::new(Node::new(Array1::zeros(0))));

        for (action, child) in &self.children {
            let mut q_value = 0.0;
            if child.borrow().visit_count > 0 {
                q_value = child.borrow().reward + child.borrow().value;
            }

            let u_value = c_puct * self.policy.as_ref().unwrap()[*action] * (self.visit_count as f64).sqrt() / (1.0 + child.borrow().visit_count as f64);
            let action_value = q_value + u_value;

            if action_value > best_value {
                best_value = action_value;
                best_action = *action;
                best_child = Rc::clone(child);
            }
        }

        (best_action, best_child)
    }

    fn update_value(&mut self, value: f64) {
        self.value = (self.visit_count as f64 * self.value + value) / (self.visit_count as f64 + 1.0);
        self.visit_count += 1;
    }

    fn expanded(&self) -> bool {
        !self.children.is_empty()
    }
}

struct MuZeroDynamics {
    // Implement the dynamics model here
}

struct MuZeroAgent {
    dynamics: MuZeroDynamics,
    prediction: MuZeroPrediction,
}

impl MuZeroAgent {
    fn new(cjepa_model: &CJEPAModel) -> Self {
        MuZeroAgent {
            dynamics: MuZeroDynamics::new(cjepa_model),
            prediction: MuZeroPrediction::new(cjepa_model),
        }
    }

    fn forward(&self, state: Array1<f64>, action: usize) -> (Array1<f64>, f64, Array1<f64>, f64) {
        let (next_state_embedding, reward) = self.dynamics.forward(state, action);
        let (policy_logits, value) = self.prediction.forward(next_state_embedding);
        (next_state_embedding, reward, policy_logits, value)
    }
}

struct MuZeroPrediction {
    // Implement the prediction model here
}

fn muzero_search(agent: &MuZeroAgent, state: Array1<f64>, num_simulations: u32) -> usize {
    let mut root = Node::new(state);
    for _ in 0..num_simulations {
        let mut node = Rc::new(RefCell::new(root.clone()));
        let mut search_path = vec![Rc::clone(&node)];

        while node.borrow().expanded() {
            let (action, child_node) = node.borrow().select_child(1.0);
            node = child_node;
            search_path.push(Rc::clone(&node));
        }

        let parent = search_path[search_path.len() - 2].clone();
        let state = parent.borrow().state.clone();
        let action = node.borrow().action.unwrap();

        // Expand the node using the prediction function
        let (next_state_embedding, reward, policy_logits, value) = agent.forward(state, action);
        node.borrow_mut().expand(policy_logits, reward, value);

        // Backpropagate the value estimates
        for node in search_path.iter().rev() {
            node.borrow_mut().update_value(value);
        }
    }

    // Select the action with the highest visit count
    root.children.iter().max_by_key(|(_action, child)| child.borrow().visit_count).map(|(action, _)| *action).unwrap()
}

fn muzero_training(agent: &mut MuZeroAgent, replay_buffer: &ReplayBuffer, num_epochs: u32, batch_size: usize, lr: f64) {
    let mut optimizer = Adam::new(lr);

    for _ in 0..num_epochs {
        for batch in replay_buffer.sample_batch(batch_size) {
            let (obs_batch, action_batch, reward_batch, done_batch, policy_batch, value_batch) = batch;

            // Compute model predictions
            let (_, rewards, policy_logits, values) = agent.forward(obs_batch, action_batch);

            // Compute losses
            let policy_loss = cross_entropy_loss(&policy_logits, &policy_batch);
            let value_loss = mse_loss(&values, &value_batch);
            let reward_loss = mse_loss(&rewards, &reward_batch);

            let total_loss = policy_loss + value_loss + reward_loss;

            // Update model parameters
            optimizer.zero_grad();
            total_loss.backward();
            optimizer.step();
        }
    }
}